factor cursor helpers into own file
